+++ gladanalysis/routes/api/v1/ms_router.py
@@ -7,27 +7,10 @@ import requests
 
 from . import endpoints
 from gladanalysis.services import GeostoreService
+from gladanalysis.services import DateService
 from gladanalysis.responders import ErrorResponder
 from gladanalysis.utils.http import request_to_microservice
 from gladanalysis.validators import validate_geostore, validate_period, validate_admin, validate_use, validate_wdpa
-
-#Testing branch history
-
-def date_to_julian_day(period):
-    #Helper function to transform dates
-    try:
-	    period_from = period.split(',')[0]
-	    period_to = period.split(',')[1]
-	    date_obj_from = datetime.datetime.strptime(period_from, '%Y-%m-%d')
-	    date_obj_to = datetime.datetime.strptime(period_to, '%Y-%m-%d')
-	    time_tuple_from = date_obj_from.timetuple()
-	    time_tuple_to = date_obj_to.timetuple()
-	    logging.info(time_tuple_from.tm_year)
-	    logging.info(time_tuple_to.tm_year)
-	    return str(time_tuple_from.tm_year), str(time_tuple_from.tm_yday), str(time_tuple_to.tm_year), str(time_tuple_to.tm_yday)
-
-    except ValueError:
-        return None, None
 
 def format_glad_sql(from_year, from_date, to_year, to_date, iso=None, state=None, dist=None):
 
@@ -185,7 +168,7 @@ def query_glad():
     period = request.args.get('period', None)
     conf = request.args.get('gladConfirmOnly', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
     sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date)
@@ -220,7 +203,7 @@ def query_terrai():
     geostore = request.args.get('geostore', None)
     period = request.args.get('period', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #create conditions that issue correct sql
     sql, download_sql = format_terrai_sql(from_year, from_date, to_year, to_date)
@@ -248,7 +231,7 @@ def glad_dist(iso_code, admin_id, dist_id):
     period = request.args.get('period', None)
     conf = request.args.get('gladConfirmOnly', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
     sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date, iso_code, admin_id, dist_id)
@@ -280,7 +263,7 @@ def glad_admin(iso_code, admin_id):
     period = request.args.get('period', None)
     conf = request.args.get('gladConfirmOnly', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
     sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date, iso_code, admin_id)
@@ -313,7 +296,7 @@ def glad_country(iso_code):
     period = request.args.get('period', None)
     conf = request.args.get('gladConfirmOnly', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
     sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date, iso_code)
@@ -343,7 +326,7 @@ def terrai_admin(iso_code, admin_id):
 
     period = request.args.get('period', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send dates to sql formatter
     sql, download_sql = format_terrai_sql(from_year, from_date, to_year, to_date, iso_code, admin_id)
@@ -368,7 +351,7 @@ def terrai_dist(iso_code, admin_id, dist_id):
 
     period = request.args.get('period', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send dates to sql formatter
     sql, download_sql = format_terrai_sql(from_year, from_date, to_year, to_date, iso_code, admin_id, dist_id)
@@ -394,7 +377,7 @@ def terrai_country(iso_code):
 
     period = request.args.get('period', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send dates to sql formatter
     sql, download_sql = format_terrai_sql(from_year, from_date, to_year, to_date, iso_code)
@@ -421,7 +404,7 @@ def glad_use(use_type, use_id):
     period = request.args.get('period', None)
     conf = request.args.get('gladConfirmOnly', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
     sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date)
@@ -451,7 +434,7 @@ def terrai_use(use_type, use_id):
 
     period = request.args.get('period', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
     sql, download_sql = format_terrai_sql(from_year, from_date, to_year, to_date)
@@ -477,7 +460,7 @@ def glad_wdpa(wdpa_id):
     period = request.args.get('period', None)
     conf = request.args.get('gladConfirmOnly', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
     sql, download_sql = format_glad_sql(from_year, from_date, to_year, to_date)
@@ -507,7 +490,7 @@ def terrai_wdpa(wdpa_id):
 
     period = request.args.get('period', None)
 
-    from_year, from_date, to_year, to_date = date_to_julian_day(period)
+    from_year, from_date, to_year, to_date = DateService.date_to_julian_day(period)
 
     #send to sql formatter function
     sql, download_sql = format_terrai_sql(from_year, from_date, to_year, to_date)
