+++ gladanalysis/services/__init__.py
@@ -5,3 +5,4 @@ from __future__ import division
 from __future__ import print_function
 
 from gladanalysis.services.geostore_service import GeostoreService
+from gladanalysis.services.date_service import DateService
